refactor: extract account creation from external identity logic (#777)
Extracts the account creation from external identities into a separate
function which can be refactored further on.

The extracted function intentionally uses a minimal number of
transformations to keep it bugs free. This is why the `terr` error
variable is still used, even though it should be renamed to `err`. Also
why the `errReturnNil` mechanism is being used. This will be refactored
in follow up PRs.