feat: add idle db connection options (duration, count, healthcheck period) (#811)
Adds the following idle DB connection options:

`GOTRUE_DB_MAX_IDLE_POOL_SIZE` - maximum number of idle connections
(defaults to 2, if not set or 0)
`GOTRUE_DB_CONN_MAX_LIFETIME` - maximum duration of a connection
(defaults to unlimited, if not set or 0)
`GOTRUE_DB_CONN_MAX_IDLE_TIME` - maximum idle duration of a connection
(defaults to unlimited, if not set or 0)
`GOTRUE_DB_HEALTH_CHECK_PERIOD` - duration between health check (pings)
on the connection when idle (defaults to 1 minute, if not set or 0)

The durations are expressed as Go durations. So `1s` for 1 second, or
`1m` for one minute.