fix: account linking logic (#990)
## What kind of change does this PR introduce?
* In older versions, there can be instances where a user exists without
an identity. When we refactored the account linking logic in
https://github.com/supabase/gotrue/releases/tag/v2.35.0, we refactored
the account linking logic to check if there is an existing identity with
the same email. If a user doesn't have any identities, an oauth sign-in
(which should automatically link) will fail since the decision made
would be to `CreateAccount`.

## Steps to test
1. sign up a user with magiclink and confirm the email
2. ssh into the db and delete the user's identity
3. attempt to use the same email to sign in with google (fails as
expected)
4. build gotrue with these changes
5. repeat step (3) and observe that the oauth sign-in works now

Steps 1-2 attempts to recreate the scenario where a user exists without
an identity.