feat: PKCE magic link (#1016)
## What kind of change does this PR introduce?

This PR adds PKCE to magic link and sign up verification types. The
remaining flows - `email_change` and `recovery` will come soon. The
`invite` type is subject to discussion.


We introduce a new column `authentication_method` to assis with tracking
of `amr` claims when we exchange an Auth Code for a session and trigger
the generation of an access token.


As an overview, the user will now receive a token `pkce_<token>` in the
email when making a `/otp` request to PKCE - this will create a flow
state containing the relevant details of the flow state.

The user will then call `/verify` using this token and GoTrue will
search for the latest flow state for a user and attach a user's ID to
that state.

TODO: 
- [x] Write integration test
- [x] manually test against test app

---------

Co-authored-by: joel@joellee.org <joel@joellee.org>
Co-authored-by: Stojan Dimitrovski <sdimitrovski@gmail.com>