feat: add pkce to email_change routes (#1082)
## What kind of change does this PR introduce?

Add PKCE to email Change routes

## What is the current behavior?

No PKCE on email change routes

## What is the new behavior?

PKCE on email change routes 


## Additional context

There's an additional AMR claim known as `email_change` I'm not sure
whether we want to have a special claim for this given that
`email_change` is not typically classed as a login method. The other
option would be to use the Magic Link AMR claim instead. Let me know if
anyone has a preference here.

Should be tested with:
https://github.com/supabase/gotrue-js/pull/661/files


TODOs:
- [x] Tests need to be written for the PKCE path

---------

Co-authored-by: joel@joellee.org <joel@joellee.org>