fix: add guard check in case factor, session, or user are missing (#1099)
Aims to prevent the existing issue where the session seems to be lost
and a null pointer execption is raised.

HS ID: 1575266879

The root cause is still unidentified and we have only been able to
reproduce once. Hoping that with the guard check we can flag more
instances. Last recorded occurrence was in April


We will follow up with the user to see if there are any repeat
occurences

---------

Co-authored-by: joel@joellee.org <joel@joellee.org>