feat: complete OIDC support for Apple and Google providers (#1108)
Previously OIDC sign in (i.e. sign-in using an ID token) for Apple,
Google and a few other providers was not properly supported. There was
no account linking available, and there were a few security issues found
with the implementation.

This PR attempts to resolve all of the issues, specifically targeting
Apple and Google providers, which enables native Sign in with Apple and
Google with mobile or desktop apps. Furthermore, this PR paves the way
towards SSO with OIDC support.

Basically, the whole `POST /token?grant_type=id_token` endpoint is
refactored to use the central `createAccountFromExternalIdentity` method
which supports both regular and SSO accounts with automatic account
linking.

For both Apple and Google flows, the important thing to realize is that
their OAuth2 flows are in-fact OIDC authentication flows. The Apple
OAuth2 flow already used the Apple OIDC ID token to extract user
information. The Google OAuth2 flow is refactored to use the OIDC ID
token when available (appears to be always) or fall back to the previous
implementation.

Since it does not matter whether the flow is OAuth2 or OIDC, automatic
account linking can take place.

The remaining OIDC supported providers -- Azure, Facebook, Keycloak --
remain supported though with upgraded account linking support; however
their implementations are best-effort at this point. Furthermore, the
Keycloak implementation should be deprecated as it's actually solving a
SSO-with-OIDC problem.