feat: serialized access to session in `refresh_token` grant (#1190)
Improves the `refresh_token` grant flow by serializing access over the
session ID, so that a session cannot concurrently be refreshed.

It achieves this by adding a boolean `forUpdate` parameter to
`models.FindSessionByID()` and `models.FindUserWithRefreshToken()`. This
in turn uses a [`SELECT ... FOR
UPDATE`](https://www.postgresql.org/docs/current/sql-select.html#SQL-FOR-UPDATE-SHARE)
query that locks the row from use with other flows that select it with a
`FOR UPDATE` clause.