feat: return `expires_at` in addition to `expires_in` (#1183)
There are cases where the response for an access and refresh token takes
more than 1 second. In such cases, `gotrue-js` will record the *expiry
time* as the time it received the response + `expires_in`. However, this
is not correct because the access token is likely to have already
expired by the recorded time.

With this change, `gotrue-js` can just use an `expires_at` value
instead. `expires_in` is still sent for backward compatibility.